requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
scikit-learn
torch
httpx
python-dotenv
geopy
fastapi[all]